+++ spec/services/pdf_converter_spec.rb
@@ -4,17 +4,18 @@ describe PDFConverter do
   let(:user) { create :user }
   let(:journal) { create :journal, pdf_css: 'body { background-color: red; }' }
   let(:paper) { create :paper, creator: create(:user), journal: journal }
+  let(:converter) { PDFConverter.new(paper, user) }
 
   describe '#convert' do
     it 'uses PDFKit to generate PDF' do
       expect(PDFKit).to receive_message_chain(:new, :to_pdf)
-      PDFConverter.new(paper, user).convert
+      converter.convert
     end
   end
 
   describe '.pdf_html' do
     let(:doc) { Nokogiri::HTML(pdf_html) }
-    let(:pdf_html) { PDFConverter.new(paper, user).pdf_html }
+    let(:pdf_html) { converter.pdf_html }
 
     after { expect(doc.errors.length).to be 0 }
 
@@ -40,14 +41,12 @@ describe PDFConverter do
 
     context 'when paper has supporting information files' do
       let(:file) do
-        with_aws_cassette 'supporting_info_files_controller' do
-          paper.supporting_information_files
-            .create! attachment: ::File.open('spec/fixtures/yeti.tiff')
-        end
+        paper.supporting_information_files
+          .create! attachment: ::File.open('spec/fixtures/yeti.tiff')
       end
 
       it 'has supporting information' do
-        expect(file)
+        expect(file).to be_truthy
         expect(paper.supporting_information_files.length).to be 1
         expect(doc.css('#si_header').count).to be 1
         expect(doc.css("img#si_preview_#{file.id}").count).to be 1
@@ -59,7 +58,7 @@ describe PDFConverter do
         expect(file)
         expect(doc.css('.si_preview').count).to be 1
         expect(doc.css('.si_preview').first['src'])
-          .to have_s3_url(file.attachment.url)
+          .to have_s3_url(file.attachment.url(:preview))
       end
 
       it 'the si_link urls are non-expiring proxy urls' do
@@ -67,6 +66,40 @@ describe PDFConverter do
         expect(doc.css('.si_link').count).to be 1
         expect(doc.css('.si_link').first['href'])
           .to eq file.non_expiring_proxy_url(only_path: false)
+      end
+    end
+
+    context 'when paper has figures' do
+      before do
+        paper.figures
+          .create attachment: File.open('spec/fixtures/yeti.tiff'),
+                  status: 'done'
+      end
+
+      it 'replaces img src urls (which are normally proxied) with resolveable
+        urls' do
+        # since pdfs maker apparently cant resolve proxy url for img.src
+        # this will completed in
+        # https://developer.plos.org/jira/browse/APERTA-5741
+
+        figure = paper.figures.first
+        paper.body = "<img id='figure_#{figure.id}' src='foo'/>"
+
+        img = doc.css("img#figure_#{figure.id}").first
+        expect(img['src']).to have_s3_url(figure.attachment.url(:detail))
+      end
+
+      it 'works with orphan figures' do
+        # add another figure
+        paper.figures
+          .create attachment: File.open('spec/fixtures/yeti.tiff'),
+                  status: 'done'
+        fig1, fig2 = paper.figures
+        paper.body = "<img id='figure_#{fig1.id}' src='foo'/>"
+        expect(converter.orphan_figures).to eq([fig2])
+
+        expect(doc.css("img#figure_#{fig2.id}").first['src']).to \
+          eq(fig2.attachment.url(:preview))
       end
     end
   end